requirements.txt
django==3.0.8
djangorestframework==3.11.0
pillow==7.2.0
gunicorn==20.1.0 
psycopg2==2.8.6
dj-static==0.0.6
static3==0.7.0
django-cors-headers==3.7.0